modeloAI/requirements.txt
flask
tensorflow==2.18.0
pillow
